React input
